Move Authentication Entity -> Model
