+++ src/OpenConext/EngineBlock/Authentication/Dto/Consent.php
@@ -5,7 +5,7 @@ namespace OpenConext\EngineBlock\Authentication\Dto;
 use DateTime;
 use OpenConext\Component\EngineBlockMetadata\ContactPerson;
 use OpenConext\Component\EngineBlockMetadata\Entity\ServiceProvider;
-use OpenConext\EngineBlock\Authentication\Entity\Consent as ConsentEntity;
+use OpenConext\EngineBlock\Authentication\Model\Consent as ConsentEntity;
 
 final class Consent
 {
