+++ src/OpenConext/EngineBlock/Service/ConsentService.php
@@ -6,7 +6,7 @@ use Doctrine\DBAL\DBALException;
 use OpenConext\EngineBlock\Authentication\Dto\Consent;
 use OpenConext\EngineBlock\Authentication\Dto\ConsentList;
 use OpenConext\EngineBlock\Exception\RuntimeException;
-use OpenConext\EngineBlock\Authentication\Entity\Consent as ConsentEntity;
+use OpenConext\EngineBlock\Authentication\Model\Consent as ConsentEntity;
 use OpenConext\EngineBlock\Authentication\Repository\ConsentRepository;
 use OpenConext\Value\Saml\EntityId;
 use Psr\Log\LoggerInterface;
